requirements.txt
beautifulsoup4==4.9.3
bleach==3.2.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
Django==2.2.17
django-appconf==1.0.4
django-avatar==5.0.0
django-bootstrap4==2.3.1
django-comments-dab==2.0.0
django-comments-xtd==2.8.1
django-contrib-comments==1.9.2
django-downtime==1.1.2
django-flag-app==0.1.1
django-jsonfield==1.4.1
django-notify-x==0.1.9
django-photologue==3.13
django-recaptcha==2.0.6
django-sortedm2m==3.0.2
django-tinymce==3.1.0
djangorestframework==3.12.2
docutils==0.16
ExifRead==2.3.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.0.0
lxml==4.6.1
mysqlclient==2.0.1
numpy==1.19.4
packaging==20.4
pandas==1.1.4
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
requests==2.24.0
scipy==1.5.4
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
timeago==1.0.14
urllib3==1.25.11
webencodings==0.5.1
wincertstore==0.2
zipp==3.4.0
